+++ src/api/video.ts
@@ -1,6 +1,6 @@
 import { RevError } from '../rev-error';
 import type { RevClient } from '../rev-client';
-import { Video, Rev } from '../types';
+import { Video, Rev, Admin } from '../types';
 import { SearchRequest } from '../utils/request-utils';
 import { videoReportAPI } from './video-report-request';
 import { videoDownloadAPI } from './video-download';
@@ -8,6 +8,15 @@ import { videoDownloadAPI } from './video-download';
 type VideoSearchDetailedItem = Video.SearchHit & (Video.Details | { error?: Error });
 
 export default function videoAPIFactory(rev: RevClient) {
+    /** get list of comments on a video
+         * set showAll param to true to include un-redacted values of comments (admin only)
+         */
+    function comments(videoId: string): Promise<Video.Comment[]>;
+    function comments(videoId: string, showAll: true): Promise<Video.Comment.Unredacted[]>;
+    async function comments(videoId: string, showAll: boolean = false): Promise<Video.Comment[] | Video.Comment.Unredacted[]> {
+        const response = await rev.get<Video.Comment.ListResponse>(`/api/v2/videos/${videoId}/comments`, showAll ? { showAll: 'true' } : undefined);
+        return response.comments;
+    }
     const videoAPI = {
         /**
          * This is an example of using the video Patch API to only update a single field
@@ -16,6 +25,18 @@ export default function videoAPIFactory(rev: RevClient) {
          */
         async setTitle(videoId: string, title: string) {
             const payload = [{ op: 'add', path: '/Title', value: title }];
+            await rev.patch(`/api/v2/videos/${videoId}`, payload);
+        },
+        /**
+         * Use the Patch API to update a single Custom Field.
+         * @param videoId - id of video to update
+         * @param customField - the custom field object (with id and value)
+         */
+        async setCustomField(videoId: string, customField: Pick<Admin.CustomField, 'id' | 'value'>) {
+            const payload = [
+                { op: 'remove', path: '/customFields', value: customField.id },
+                { op: 'add', path: '/customFields/-', value: customField }
+            ];
             await rev.patch(`/api/v2/videos/${videoId}`, payload);
         },
         /**
@@ -28,14 +49,10 @@ export default function videoAPIFactory(rev: RevClient) {
         async details(videoId: string): Promise<Video.Details> {
             return rev.get(`/api/v2/videos/${videoId}/details`);
         },
-        /** get list of comments on a video */
-        async comments(videoId: string): Promise<Video.Comment[]> {
-            const response = await rev.get<Video.Comment.ListResponse>(`/api/v2/videos/${videoId}/comments`);
-            return response.comments;
-        },
+        comments,
         async chapters(videoId: string): Promise<Video.Chapter[]> {
             try {
-                const {chapters} = await rev.get<{chapters: Video.Chapter[]}>(`/api/v2/videos/${videoId}/comments`);
+                const {chapters} = await rev.get<{chapters: Video.Chapter[]}>(`/api/v2/videos/${videoId}/chapters`);
                 return chapters;
             } catch (err) {
                 // if no chapters then this api returns a 400 response
@@ -114,7 +131,10 @@ export default function videoAPIFactory(rev: RevClient) {
             return video;
         },
         ...videoDownloadAPI(rev),
-        ...videoReportAPI(rev)
+        ...videoReportAPI(rev),
+        async trim(videoId: string, removedSegments: Array<{ start: string, end: string }>) {
+            return rev.post(`/api/v2/videos/${videoId}/trim`, removedSegments);
+        }
     };
     return videoAPI;
 }